requirements.txt
# MindX Minimal Dependencies - Essential Only
# This is a minimal set of dependencies to get MindX working

# Core async and HTTP
aiohttp>=3.8.0
aiofiles>=23.0.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.20.0

# Data processing
pydantic>=2.0.0

# AI/ML Libraries
openai>=1.0.0
anthropic>=0.30.0
google-generativeai>=0.8.0
mistralai>=1.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
requests>=2.30.0

# Logging
colorama>=0.4.0

# Testing
pytest>=7.0.0
